Jc error handling

+++ SceneSherpa/Controllers/ReviewsController.cs
@@ -22,21 +22,33 @@ namespace SceneSherpa.Controllers
         }
 
         [Route("media/{id:int}/reviews/{reviewId:int}")]
-        public IActionResult Edit(int reviewId)
+        public IActionResult Edit(int? reviewId)
         {
+            if (reviewId == null)
+            {
+                return BadRequest();
+            }
             ViewBag.MediaList = _context.Media.ToList();
             var review = _context.Reviews.Where(r => r.Id == reviewId).Include(r => r.Media).First();
-
+            if (review == null)
+            {
+                return NotFound();
+            }
             ViewData["CurrentUserIdUsername"] = Request.Cookies["CurrentUserIdUsername"];
+            
             return View(review);
         }
 
         [HttpPost]
         [Route("media/{mediaId:int}/reviews/{reviewId:int}")]
-        public IActionResult Update(int mediaId, Review review, int reviewId)
+        public IActionResult Update(int mediaId, Review review, int? reviewId)
         {
+            if (reviewId == null)
+            {
+                return BadRequest();
+            }
             ViewBag.MediaList = _context.Media.ToList();
-            review.Id = reviewId;
+            review.Id = (int)reviewId;
             review.UpdatedAt = DateTime.Now.ToUniversalTime();
             review.Content = Markdown.Parse(review.Content);
             _context.Reviews.Update(review);
@@ -47,8 +59,12 @@ namespace SceneSherpa.Controllers
 
         [HttpPost]
         [Route("/Media/{mediaId:int}/Reviews")]
-        public IActionResult Create(int mediaId, Review review)
+        public IActionResult Create(int? mediaId, Review review)
         {
+            if (mediaId == null)
+            {
+                return BadRequest();
+            }
             //Getting Currently Logged in User
             var currentUsername = Request.Cookies["CurrentUserIdUsername"].Split()[1];
             var currentUserId = _context.Users
@@ -61,6 +77,10 @@ namespace SceneSherpa.Controllers
                 .Where(m => m.Id == mediaId)
                 .Include(m => m.Reviews)
                 .First();
+            if (media == null)
+            {
+                return NotFound();
+            }
             review.CreatedAt = DateTime.Now.ToUniversalTime();
             review.Content = Markdown.Parse(review.Content);
             media.Reviews.Add(review);
@@ -69,20 +89,35 @@ namespace SceneSherpa.Controllers
         }
         
         [Route("/Media/{mediaId:int}/Reviews/New")]
-        public IActionResult New(int mediaId)
+        public IActionResult New(int? mediaId)
         {
+            if (mediaId == null)
+            {
+                return BadRequest();
+            }
             ViewBag.MediaList = _context.Media.ToList();
             var media = _context.Media.Where(m => m.Id == mediaId).Include(m => m.Reviews).First();
-
+            if (media == null)
+            {
+                return NotFound();
+            }
             ViewData["CurrentUserIdUsername"] = Request.Cookies["CurrentUserIdUsername"];
             return View(media);
         }
 
         [HttpPost]
         [Route("media/{mediaId:int}/reviews/delete/{reviewId:int}")]
-        public IActionResult Delete(int reviewId, int mediaId)
+        public IActionResult Delete(int? reviewId, int mediaId)
         {
+            if (reviewId == null)
+            {
+                return BadRequest();
+            }
             var review = _context.Reviews.Find(reviewId);
+            if (review == null)
+            {
+                return NotFound();
+            }
             _context.Reviews.Remove(review);
             _context.SaveChanges();
